requirements.txt
cycler==0.12.1
fonttools==4.43.1
h5py==3.10.0
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.26.4
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
mypy-extensions==1.0.0
